[option] Adds options to allow operator tunes cluster mode based on network and usage
- gossip_interval value
  specify the interval between sending messages that need to be gossiped that haven't been able to piggyback on probing messages in millisecond (default 200)
- gossip_packet_size value
  specify the size of a gossip packet in byte (will be for UDP packet), it depends on your network's MTU (default 4000)

+++ src/cluster/config.go
@@ -33,7 +33,7 @@ type Config struct {
 
 	GossipNodes, IndirectCheckNodes uint
 
-	UDPBufferSize int
+	UDPBufferSize uint16
 
 	MessageSendTimeout, FailedMemberReconnectTimeout, MemberLeftRecordTimeout time.Duration
 	RecentMemberOperationTimeout                                              time.Duration
@@ -126,7 +126,7 @@ func createMemberListConfig(conf *Config, eventDelegate memberlist.EventDelegate
 		DisableTcpPings:         false,
 		DNSConfigPath:           "/etc/resolv.conf",
 		HandoffQueueDepth:       1024,
-		UDPBufferSize:           udpBufferSize,
+		UDPBufferSize:           int(udpBufferSize),
 		RetransmitMult:          gossipRetransmitMult,
 		SuspicionMult:           memberSuspicionMult,
 		SuspicionMaxTimeoutMult: memberSuspicionMaxTimeoutMult,
